+++ djangocms_version_locking/test_utils/polls/admin.py
@@ -1,10 +1,12 @@
 from django.contrib import admin
 
+from djangocms_version_locking.admin import VersionLockContentAdminMixin
+
 from .models import Answer, Poll, PollContent
 
 
 @admin.register(PollContent)
-class PollContentAdmin(admin.ModelAdmin):
+class PollContentAdmin(VersionLockContentAdminMixin, admin.ModelAdmin):
     pass